requirements.txt
pytest
pytest-html
pytest-xdist
selenium
requests
allure-pytest
jsonschema
